# gh-101196: Make isdir/isfile/exists faster on Windows