--- a/Lib/ntpath.py
+++ b/Lib/ntpath.py
@@ -852,11 +852,12 @@ def commonpath(paths):
 
 
 try:
-    # The isdir(), isfile(), and exists() implementations in genericpath use
-    # os.stat(). This is overkill on Windows. Use simpler builtin functions
-    # if they are available.
+    # The isdir(), isfile(), islink() and exists() implementations in
+    # genericpath use os.stat(). This is overkill on Windows. Use simpler
+    # builtin functions if they are available.
     from nt import _isdir as isdir
     from nt import _isfile as isfile
+    from nt import _islink as islink
     from nt import _exists as exists
 except ImportError:
     # Use genericpath.* as imported above