… Doc/library/os.path.rst
… Lib/test/test_ntpath.py
… Lib/test/test_os.py