@eryksun 2023-01-25
A builtin `_islink()` implementation would open the path using the flags `FILE_FLAG_OPEN_REPARSE_POINT | FILE_FLAG_BACKUP_SEMANTICS` and get the [`FileAttributeTagInfo`](https://learn.microsoft.com/en-us/windows/win32/api/winbase/ns-winbase-file_attribute_tag_info). It's a symlink if the query succeeds and `ReparseTag` is `IO_REPARSE_TAG_SYMLINK`. Note that `IO_REPARSE_TAG_MOUNT_POINT` (i.e. a junction) is intentionally excluded from `islink()`.