@eryksun 2023-01-25
A few cases that are handled by `win32_xstat_impl()` should be supported. Perhaps a common function could be implemented that factors out the initial open from `win32_xstat_impl()` and `attributes_from_dir()` -- e.g. `win32_stat_open(const wchar_t *path, BOOL traverse, WIN32_FIND_DATAW *find_data, BOOL *find_data_updated)`. Or perhaps simply fall back on `win32_xstat_impl()` if `CreateFileW()` fails with one of the error codes discussed below. 

---

* `ERROR_ACCESS_DENIED` and `ERROR_SHARING_VIOLATION`: Fall back on using `FindFirstFileW()`. Internally, this opens the parent directory via `NtOpenFile()` and queries a directory entry for the name using `NtQueryDirectoryFileEx()`. The query must fail or return false if the directory entry is a reparse point, unless it's a name-surrogate reparse point (e.g. a symlink or junction) that was supposed to be opened. Here are some cases where this can occur:
    * The file security may grant no access to the caller. Microsoft's filesystems implement a policy that always grants `FILE_READ_ATTRIBUTES` access if the caller is granted `FILE_READ_DATA` access to the parent directory. However, `CreateFileW()` also requests `SYNCHRONIZE` access.
    * No access is granted to a file that's mapped as a system paging file.
    * No access is granted to a file or directory that has been deleted but is not unlinked, which is the case if there are existing opens and POSIX delete isn't supported by 
        * the OS version (e.g. Windows 8), 
        * the API function (e.g. `RemoveDirectoryW`), 
        * or the filesystem driver (e.g. exFAT and FAT32).
* `ERROR_CANT_ACCESS_FILE`: Traversing a reparse point in a path may not be supported because it was created by software on another system (e.g. on a portable drive), or the driver/service that supports the reparse point is no longer installed or running, or it was simply never intended for automatic traversal (e.g. `IO_REPARSE_TAG_APPEXECLINK`). In this case, the open should be retried with the flag `FILE_FLAG_OPEN_REPARSE_POINT`. If it turns out to be a name-surrogate reparse point, the query should be failed with the original error code, `ERROR_CANT_ACCESS_FILE`, as a broken symlink(ish) file. 
* `ERROR_INVALID_PARAMETER`: A device open may require `GENERIC_READ` access. For example, opening "\\\\.\\CON" (i.e. "\\Device\\ConDrv\\Console") requires either `GENERIC_READ` or `GENERIC_WRITE` access. The open has to be retried with the required `GENERIC_READ` access.