@eryksun 2023-01-25
> > Or perhaps simply fall back on win32_xstat_impl() if CreateFileW() fails with one of the error codes discussed below.
> 
> My initial implementation from this morning did that, which I like because it doesn't repeat what's clearly battle-tested code, but it made things _slower_ in the event of symlinks.

There's should be an issued with speed for common cases as long as you only fall back on `win32_xstat_impl()` for the discussed error codes:

* `ERROR_ACCESS_DENIED`
* `ERROR_SHARING_VIOLATION`
* `ERROR_CANT_ACCESS_FILE`
* `ERROR_INVALID_PARAMETER`