@mdboom 2023-01-26
@eryksun: Thanks for all of your help and pointers.

I've implemented support for these corner cases by calling `win32_stat` under the hood, which is not the most efficient but avoids duplicating the complex logic in `win32_xstat_impl`.

I'm a little stuck on writing unit tests for them.  I was able to cover `ERROR_ACCESS_DENIED` by extending an existing test for `stat` to also ensure that `isfile` matches it.

I'm not sure how to recreate the case you describe in `ERROR_CANT_ACCESS_FILE`.

For `ERROR_INVALID_PARAMETER`, I tried `isfile(r"\\.\CON")` and `isdir(r"\\.CON")`, but I get the same result (`False`) with or without this error handling as well as on `main`.  So maybe I need to somehow create or have `CON`?