@eryksun 2023-01-26
Maybe also add `nt._isjunction()` for the new `ntpath.isjunction()` test in 3.12. It's like `islink()`, except it checks for the tag `IO_REPARSE_TAG_MOUNT_POINT`. In the `LSTAT` fallback, it would check `st.st_reparse_tag == IO_REPARSE_TAG_MOUNT_POINT`.